requirements.txt
# Core dependencies for Federated Learning with Fog Computing
torch>=2.0.0,<3.0.0          # Deep learning framework
torchvision>=0.15.0,<0.18.0  # Computer vision extensions for PyTorch
numpy>=1.21.0,<2.0.0         # Fundamental package for array computing - compatible with PyTorch
scikit-learn>=1.0.0,<2.0.0   # Machine learning library
scipy>=1.7.0,<2.0.0          # Scientific computing library
flwr>=1.0.0,<2.0.0           # Flower federated learning framework
paho-mqtt>=2.0.0,<3.0.0      # MQTT client library
pyyaml>=6.0.0,<7.0.0         # YAML parsing for federated configs

# Data processing and time series
pandas>=1.3.0,<3.0.0         # Data manipulation and analysis
tslearn>=0.5.0,<1.0.0        # Time series machine learning

# Visualization and reporting
matplotlib>=3.5.0,<4.0.0     # Plotting library
seaborn>=0.11.0,<1.0.0       # Statistical data visualization
tabulate>=0.9.0,<1.0.0       # Pretty-print tabular data

# Development and testing (optional)
pytest>=7.0.0,<8.0.0         # Testing framework
pytest-asyncio>=0.21.0,<2.0.0 # Async testing support
pytest-mock>=3.10.0,<4.0.0   # Mocking library for tests
pytest-cov>=4.0.0,<8.0.0     # Coverage reporting
black>=22.0.0,<26.0.0       # Code formatter
isort>=5.10.0,<7.0.0        # Import sorter
flake8>=4.0.0,<8.0.0        # Linter
mypy>=1.0.0,<2.0.0          # Type checker
ruff>=0.1.0,<1.0.0          # Fast Python linter
pre-commit>=2.17.0,<5.0.0   # Git hooks framework
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-exporter-otlp>=1.24.0,<2.0.0
